\name{shrinksv-package}
\alias{shrinksv-package}
\alias{shrinksv}
\docType{package}
\title{
\packageTitle{shrinksv}
}
\description{
\packageDescription{shrinksv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shrinksv}
\packageIndices{shrinksv}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shrinksv}

Maintainer: \packageMaintainer{shrinksv}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
