% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_c_main.R
\name{sv_c_main}
\alias{sv_c_main}
\title{Title}
\usage{
sv_c_main(
  y,
  draws,
  burnin,
  thinpara,
  prior_spec,
  startpara,
  startlatent,
  expert
)
}
\arguments{
\item{y}{}

\item{draws}{}

\item{burnin}{}

\item{thinpara}{}

\item{prior_spec}{}

\item{startpara}{}

\item{startlatent}{}

\item{expert}{}
}
\value{

}
\description{
Title
}
